+++ api/v1alpha1/userconfig/service/opensearch/opensearch.go
@@ -28,6 +28,9 @@ type AzureMigration struct {
 	// Defines the DNS suffix for Azure Storage endpoints.
 	EndpointSuffix *string `groups:"create,update" json:"endpoint_suffix,omitempty"`
 
+	// Whether to restore aliases alongside their associated indexes. Default is true.
+	IncludeAliases *bool `groups:"create,update" json:"include_aliases,omitempty"`
+
 	// +kubebuilder:validation:Pattern=`^(\*?[a-z0-9._-]*\*?|-\*?[a-z0-9._-]*\*?)(,(\*?[a-z0-9._-]*\*?|-\*?[a-z0-9._-]*\*?))*[,]?$`
 	// A comma-delimited list of indices to restore from the snapshot. Multi-index syntax is supported. By default, a restore operation includes all data streams and indices in the snapshot. If this argument is provided, the restore operation only includes the data streams and indices that you specify.
 	Indices *string `groups:"create,update" json:"indices,omitempty"`
@@ -68,6 +71,9 @@ type GcsMigration struct {
 	// +kubebuilder:validation:Pattern=`^[^\r\n]*$`
 	// Google Cloud Storage credentials file content
 	Credentials string `groups:"create,update" json:"credentials"`
+
+	// Whether to restore aliases alongside their associated indexes. Default is true.
+	IncludeAliases *bool `groups:"create,update" json:"include_aliases,omitempty"`
 
 	// +kubebuilder:validation:Pattern=`^(\*?[a-z0-9._-]*\*?|-\*?[a-z0-9._-]*\*?)(,(\*?[a-z0-9._-]*\*?|-\*?[a-z0-9._-]*\*?))*[,]?$`
 	// A comma-delimited list of indices to restore from the snapshot. Multi-index syntax is supported. By default, a restore operation includes all data streams and indices in the snapshot. If this argument is provided, the restore operation only includes the data streams and indices that you specify.
@@ -739,6 +745,9 @@ type S3Migration struct {
 	// +kubebuilder:validation:Pattern=`^[^\r\n]*$`
 	// The S3 service endpoint to connect to. If you are using an S3-compatible service then you should set this to the service’s endpoint
 	Endpoint *string `groups:"create,update" json:"endpoint,omitempty"`
+
+	// Whether to restore aliases alongside their associated indexes. Default is true.
+	IncludeAliases *bool `groups:"create,update" json:"include_aliases,omitempty"`
 
 	// +kubebuilder:validation:Pattern=`^(\*?[a-z0-9._-]*\*?|-\*?[a-z0-9._-]*\*?)(,(\*?[a-z0-9._-]*\*?|-\*?[a-z0-9._-]*\*?))*[,]?$`
 	// A comma-delimited list of indices to restore from the snapshot. Multi-index syntax is supported. By default, a restore operation includes all data streams and indices in the snapshot. If this argument is provided, the restore operation only includes the data streams and indices that you specify.